requirements.txt
aiohttp~=3.8.4
pytest~=7.2.2
pytest-asyncio~=0.20.3
assertpy~=1.1
PyYAML~=6.0
loguru~=0.6.0
Pillow~=9.4.0

